skills/wechat-bot/requirements.txt
# 微信机器人依赖库
wxpy>=0.3.9
python-dotenv>=0.19.0